Refactor query functions and move S3 write utility to a new module
